requirements.txt
google-genai
pymupdf
streamlit
python-dotenv